refactor: deduplicate shared diff utilities and info message classifier
Extract `compute_text_diff`, `tags_eq_unordered`, and `TextDiffSummary`
from content_block/email_template into `diff::mod`. Extract
`InfoMessageClass` and `classify_info_message` into `braze::mod`,
parameterized by resource-specific not-found phrase.
