+++ src/format/fixtures.rs
@@ -4,7 +4,8 @@
 //! corresponding TableFormatter / JsonFormatter snapshot is reproducible.
 
 use crate::diff::catalog::{diff_schema, CatalogItemsDiff};
-use crate::diff::content_block::{diff as diff_content_block, ContentBlockDiff, TextDiffSummary};
+use crate::diff::content_block::{diff as diff_content_block, ContentBlockDiff};
+use crate::diff::TextDiffSummary;
 use crate::diff::custom_attribute::{CustomAttributeDiff, CustomAttributeOp};
 use crate::diff::email_template::EmailTemplateDiff;
 use crate::diff::{DiffOp, DiffSummary, ResourceDiff};
